Restructure codebase: move `backend` to `backend/src`
Relocated core files and modules under `backend/src` for better project organization. Updated import paths and adjusted configuration and file handling for compatibility. Added utility methods and improved optional configurations.
