+++ backend/README.md
@@ -1,0 +1,3 @@
+# Mood Coder API Backend
+
+> Please, PLEASE set the working directory to `/backend`, if it is not, errors are bound to arise.
