+++ backend/src/ai/README.md
