+++ backend/src/ai/router.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter
-from utils import verify_key
+from ..main.utils import verify_key
 
 router = APIRouter(prefix='/chats')
 
 @router.get('/')
 async def get_chats():
+    pass
